Fixed WiFi overwriting display--wifi uses ADC2
Cut the trace on the tft featherwing that connects the pad of pin
15 to the pad of TFT_CS, then soldered a jumper wire from TFT_CS
to pin 21, then updated the code to reflect the changes.

I also discovered that since the wifi uses ADC2, that we only have a
couple pins that we can use to connect other peripherals, and I might
need to change the SD_CS from 14 to 22 or 23.  Further testing will
be required.